[Infrastructure] Splitted the genshin_service.dart into multiple small chunks
